@@ -11,8 +11,6 @@ use Symfony\Component\Security\Core\User\PasswordUpgraderInterface;
 
 /**
  * @extends ServiceEntityRepository<User>
- *
- * @implements PasswordUpgraderInterface<User>
  */
 class UserRepository extends ServiceEntityRepository implements PasswordUpgraderInterface
 {
@@ -24,14 +22,47 @@ class UserRepository extends ServiceEntityRepository implements PasswordUpgrader
     /**
      * Used to upgrade (rehash) the user's password automatically over time.
      */
-    public function upgradePassword(PasswordAuthenticatedUserInterface $user, string $hashedPassword): void
+    public function upgradePassword(PasswordAuthenticatedUserInterface $user, string $newHashedPassword): void
     {
         if (!$user instanceof User) {
-            throw new UnsupportedUserException(sprintf('Instances of "%s" are not supported.', \get_class($user)));
+            throw new UnsupportedUserException(sprintf('Instances of "%s" are not supported.', $user::class));
         }
 
-        $user->setPassword($hashedPassword);
+        $user->setPassword($newHashedPassword);
         $this->getEntityManager()->persist($user);
         $this->getEntityManager()->flush();
+    }
+
+    /**
+     * @return User[] Returns filtered and sorted users
+     */
+    public function searchAndSort(?string $query, ?string $roleValue = null, string $sortField = 'id', string $sortDirection = 'DESC'): array
+    {
+        $qb = $this->createQueryBuilder('u');
+
+        if ($query) {
+            $qb->andWhere('LOWER(u.pseudo) LIKE LOWER(:query) OR LOWER(u.email) LIKE LOWER(:query) OR LOWER(u.nom) LIKE LOWER(:query)')
+               ->setParameter('query', '%' . strtolower($query) . '%');
+        }
+
+        if ($roleValue) {
+            $roleEnum = \App\Enum\Role::tryFrom($roleValue);
+            if ($roleEnum) {
+                $qb->andWhere('u.role = :role')
+                   ->setParameter('role', $roleEnum);
+            }
+        }
+
+        // Whitelist sort fields
+        $allowedSortFields = ['id', 'nom', 'email'];
+        if (!in_array($sortField, $allowedSortFields)) {
+            $sortField = 'id';
+        }
+        
+        $sortDirection = strtoupper($sortDirection) === 'ASC' ? 'ASC' : 'DESC';
+
+        return $qb->orderBy('u.' . $sortField, $sortDirection)
+                  ->getQuery()
+                  ->getResult();
     }
 }
